Move civilian benefits question into school selection page
