--- test/edu-benefits/5490/config/schoolSelection.unit.spec.jsx
@@ -1,45 +1,0 @@
-import React from 'react';
-import { findDOMNode } from 'react-dom';
-import { expect } from 'chai';
-import sinon from 'sinon';
-import ReactTestUtils from 'react-addons-test-utils';
-
-import { DefinitionTester, submitForm } from '../../../util/schemaform-utils.jsx';
-import formConfig from '../../../../src/js/edu-benefits/5490/config/form';
-
-describe('Edu 5490 schoolSelection', () => {
-  const { schema, uiSchema } = formConfig.chapters.schoolSelection.pages.schoolSelection;
-
-  it('should render', () => {
-    const form = ReactTestUtils.renderIntoDocument(
-      <DefinitionTester
-          schema={schema}
-          data={{}}
-          definitions={formConfig.defaultDefinitions}
-          uiSchema={uiSchema}/>
-    );
-
-    const formDOM = findDOMNode(form);
-
-    const inputs = Array.from(formDOM.querySelectorAll('input, select, textarea'));
-
-    expect(inputs.length).to.equal(13);
-  });
-
-  it('should have no required inputs', () => {
-    const onSubmit = sinon.spy();
-    const form = ReactTestUtils.renderIntoDocument(
-      <DefinitionTester
-          schema={schema}
-          onSubmit={onSubmit}
-          data={{}}
-          definitions={formConfig.defaultDefinitions}
-          uiSchema={uiSchema}/>
-    );
-    const formDOM = findDOMNode(form);
-    submitForm(form);
-
-    expect(Array.from(formDOM.querySelectorAll('.usa-input-error'))).to.be.empty;
-    expect(onSubmit.called).to.be.true;
-  });
-});
